requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-socketio==5.1.1
python-dotenv==0.19.0
pytest==6.2.5
gunicorn==20.1.0
SpeechRecognition==3.8.1
google-cloud-speech==2.17.3
openai==0.27.8
pydub==0.25.1
pyaudio==0.2.13 